requirements.txt
fastapi
uvicorn
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
asyncpg
sqlalchemy
databases
python-dotenv
python-multipart
psycopg2-binary
greenlet
alembic
PyJWT